requirements.txt
PyMuPDF
Pillow
google-generativeai
python-dotenv==1.0.1